requirements-backend.txt
fastapi>=0.115.2
uvicorn[standard]>=0.35.0
python-multipart>=0.0.18
sse-starlette>=2.1.0

# Core sql_agent deps (duplicated here for single-container install)
openai>=1.0.0
anthropic>=0.40.0
sqlalchemy>=2.0.0
sqlparse>=0.5.0
pyyaml>=6.0
python-dotenv>=1.0.0
pymysql>=1.1.0
cryptography>=42.0.0
psycopg2-binary>=2.9.0
sentence-transformers>=3.0.0
langgraph>=0.2.0
pandas>=2.0.0
plotly>=5.0.0
tabulate>=0.9.0
rich>=13.0.0
